+++ src/TimeScheduler/Testing/TestScheduler.cs
@@ -11,65 +11,110 @@ namespace TimeScheduler.Testing;
 /// <remarks>
 /// Learn more at <see href="https://github.com/egil/TimeScheduler"/>.
 /// </remarks>
-public sealed partial class TestScheduler : ITimeScheduler, IDisposable
+[Obsolete("Use System.Testing.TestTimeProvider instead. This will allow you to upgrade seamlessly the new TimeProvider API that is part of .NET 8 upon release.")]
+public partial class TestScheduler : TimeProvider, ITimeScheduler, IDisposable
 {
     internal const uint MaxSupportedTimeout = 0xfffffffe;
     internal const uint UnsignedInfinite = unchecked((uint)-1);
 
     private readonly ConcurrentDictionary<FutureAction, object?> futureActions = new();
     private readonly ConditionalWeakTable<object, FutureAction> attachedFutureActions = new();
+    private bool isDisposed;
+    private DateTimeOffset utcNow;
 
     /// <summary>
-    /// Returns a DateTimeOffset representing the <see cref="TestScheduler"/>'s
-    /// current date and time.
+    /// Gets a <see cref="DateTimeOffset"/> value whose date and time are set to the current
+    /// Coordinated Universal Time (UTC) date and time and whose offset is Zero,
+    /// all according to this <see cref="TestScheduler"/>'s notion of time.
     /// </summary>
     /// <remarks>
-    /// To advance time, call <see cref="ForwardTime(TimeSpan)"/>.
+    /// To advance time, call <see cref="ForwardTime(TimeSpan)"/> or <see cref="SetUtcNow(DateTimeOffset)"/>.
     /// </remarks>
-    public DateTimeOffset UtcNow { get; private set; }
+    [Obsolete("Use GetUtcNow() instead. This will allow you to upgrade seamlessly to TimeProvider that will be part of .NET 8.")]
+    public DateTimeOffset UtcNow => utcNow;
 
     /// <summary>
     /// Creates an instance of the <see cref="TestScheduler"/> with
-    /// <see cref="UtcNow"/> set to <see cref="DateTimeOffset.UtcNow"/>.
+    /// <see cref="DateTimeOffset.UtcNow"/> being the initial value returned by <see cref="GetUtcNow()"/>.
     /// </summary>
     public TestScheduler()
+        : this(DateTimeOffset.UtcNow)
     {
-        UtcNow = DateTimeOffset.UtcNow;
     }
 
     /// <summary>
     /// Creates an instance of the <see cref="TestScheduler"/> with
-    /// <see cref="UtcNow"/> set to <paramref name="startDateTime"/>.
+    /// <paramref name="startDateTime"/> being the initial value returned by <see cref="GetUtcNow()"/>.
     /// </summary>
-    /// <param name="startDateTime">The date and time to use for <see cref="UtcNow"/> initially.</param>
+    /// <param name="startDateTime">The initial date and time <see cref="GetUtcNow()"/> will return.</param>
     public TestScheduler(DateTimeOffset startDateTime)
     {
-        UtcNow = startDateTime;
+        utcNow = startDateTime;
     }
 
     /// <summary>
-    /// Forward the date and time represented by <see cref="UtcNow"/>
-    /// by the specified <paramref name="time"/>.
+    /// Gets a <see cref="DateTimeOffset"/> value whose date and time are set to the current
+    /// Coordinated Universal Time (UTC) date and time and whose offset is Zero,
+    /// all according to this <see cref="TestScheduler"/>'s notion of time.
     /// </summary>
-    /// <param name="time">The span of time to forward <see cref="UtcNow"/> with.</param>
+    /// <remarks>
+    /// To advance time, call <see cref="ForwardTime(TimeSpan)"/> or <see cref="SetUtcNow(DateTimeOffset)"/>.
+    /// </remarks>
+    public override DateTimeOffset GetUtcNow() => utcNow;
+
+    /// <summary>
+    /// Sets the date and time returned by <see cref="GetUtcNow()"/> to <paramref name="newUtcNew"/> and triggers any
+    /// scheduled items that are waiting for time to be forwarded.
+    /// </summary>
+    /// <param name="newUtcNew">The new UtcNow time.</param>
+    /// <exception cref="ArgumentException">If <paramref name="newUtcNew"/> is less than the value returned by <see cref="GetUtcNow()"/>.</exception>
+    public virtual void SetUtcNow(DateTimeOffset newUtcNew)
+    {
+        if (newUtcNew < utcNow)
+        {
+            throw new ArgumentException("The new UtcNow must be greater than or equal to the current UtcNow.", nameof(newUtcNew));
+        }
+
+        utcNow = newUtcNew;
+        CompleteDelayedTasks();
+    }
+
+    /// <summary>
+    /// Forward the date and time represented by <see cref="GetUtcNow()"/>
+    /// by the specified <paramref name="time"/>, and triggers any
+    /// scheduled items that are waiting for time to be forwarded.
+    /// </summary>
+    /// <param name="time">The span of time to forward <see cref="GetUtcNow()"/> with.</param>
     /// <exception cref="ArgumentException">If <paramref name="time"/> is negative or zero.</exception>
-    public void ForwardTime(TimeSpan time)
+    public virtual void ForwardTime(TimeSpan time)
     {
         if (time <= TimeSpan.Zero)
         {
             throw new ArgumentException("The timespan to forward time by must be positive.", nameof(time));
         }
 
-        UtcNow = UtcNow + time;
-        CompleteDelayedTasks();
+        SetUtcNow(utcNow + time);
+    }
+
+    /// <summary>
+    /// Disposing will cancel all scheduled work items/tasks that are waiting
+    /// for time to be forwarded.
+    /// </summary>
+    public void Dispose()
+    {
+        Dispose(true);
+        GC.SuppressFinalize(this);
     }
 
     /// <summary>
     /// Disposing the scheduler will cancel all scheduled tasks that are waiting
     /// for time to be forwarded.
     /// </summary>
-    public void Dispose()
+    protected virtual void Dispose(bool disposing)
     {
+        if (isDisposed) return;
+        isDisposed = true;
+
         attachedFutureActions.Clear();
 
         foreach (var futureAction in futureActions.Keys)
@@ -84,7 +129,7 @@ public sealed partial class TestScheduler : ITimeScheduler, IDisposable
     {
         var tasksToComplete = futureActions
             .Keys
-            .Where(x => x.CompletionTime <= UtcNow)
+            .Where(x => x.CompletionTime <= utcNow)
             .OrderBy(x => x.CompletionTime);
 
         foreach (var delayedAction in tasksToComplete)