+++ test/TimeScheduler.Tests/System/Testing/TestTimeProviderDelayTests.cs
@@ -1,14 +1,13 @@
-namespace TimeScheduler.Testing;
+namespace System.Testing;
 
-[Obsolete] // marked obsolete to stop warnings related to SUT
-public class TestSchedulerDelayTests
+public class TestTimeProviderDelayTests
 {
     [Fact]
     public void Delayed_task_is_completes()
     {
         var startTime = DateTimeOffset.UtcNow;
         var future = TimeSpan.FromTicks(1);
-        using var sut = new TestScheduler(startTime);
+        using var sut = new TestTimeProvider(startTime);
         var task = sut.Delay(TimeSpan.FromTicks(1));
 
         sut.ForwardTime(future);
@@ -20,7 +19,7 @@ public class TestSchedulerDelayTests
     public void Delayed_task_is_cancelled()
     {
         using var cts = new CancellationTokenSource();
-        using var sut = new TestScheduler();
+        using var sut = new TestTimeProvider();
         var task = sut.Delay(TimeSpan.FromTicks(1), cts.Token);
 
         cts.Cancel();