+++ test/TimeScheduler.Tests/System/Testing/TestTimeProviderTests.cs
@@ -1,0 +1,26 @@
+namespace System.Testing;
+
+public class TestTimeProviderTests
+{
+    [Fact]
+    public void ForwardTime_updates_UtcNow()
+    {
+        var startTime = DateTimeOffset.UtcNow;
+        using var sut = new TestTimeProvider(startTime);
+
+        sut.ForwardTime(TimeSpan.FromTicks(1));
+
+        sut.GetUtcNow().Should().Be(startTime + TimeSpan.FromTicks(1));
+    }
+
+    [Fact]
+    public void SetUtcNow_updates_UtcNow()
+    {
+        var startTime = DateTimeOffset.UtcNow;
+        using var sut = new TestTimeProvider(startTime);
+
+        sut.SetUtcNow(startTime + TimeSpan.FromTicks(1));
+
+        sut.GetUtcNow().Should().Be(startTime + TimeSpan.FromTicks(1));
+    }
+}