+++ test/TimeScheduler.Tests/System/TimeProviderTests.cs
@@ -1,14 +1,13 @@
-namespace TimeScheduler;
+namespace System;
 
-[Obsolete] // marked obsolete to stop warnings related to SUT
-public class DefaultSchedulerTests
+public class TimeProviderTests
 {
     internal const uint MaxSupportedTimeout = 0xfffffffe;
 
     [Fact]
     public void GetUtcNow_returns()
     {
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
 
         sut.GetUtcNow().Should().BeCloseTo(
             nearbyTime: DateTimeOffset.UtcNow,
@@ -18,7 +17,7 @@ public class DefaultSchedulerTests
     [Fact]
     public async Task Delay_waits_for_n_seconds()
     {
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var timer = Stopwatch.StartNew();
 
         await sut.Delay(TimeSpan.FromSeconds(1));
@@ -32,8 +31,8 @@ public class DefaultSchedulerTests
     [Fact]
     public async Task PeriodicTimer()
     {
-        var sut = DefaultScheduler.Instance;
-        using var periodTimer = sut.PeriodicTimer(TimeSpan.FromMilliseconds(5));
+        var sut = TimeProvider.System;
+        using var periodTimer = sut.CreatePeriodicTimer(TimeSpan.FromMilliseconds(5));
 
         var result = await periodTimer.WaitForNextTickAsync(CancellationToken.None);
 
@@ -44,7 +43,7 @@ public class DefaultSchedulerTests
     public async Task PeriodicTimer_cancelled_throws()
     {
         using var cts = new CancellationTokenSource();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var timerTask = TimerTask(cts.Token);
 
         cts.Cancel();
@@ -56,7 +55,7 @@ public class DefaultSchedulerTests
 
         async Task TimerTask(CancellationToken cancellationToken)
         {
-            using var periodTimer = sut.PeriodicTimer(TimeSpan.FromSeconds(1));
+            using var periodTimer = sut.CreatePeriodicTimer(TimeSpan.FromSeconds(1));
             await periodTimer.WaitForNextTickAsync(cancellationToken);
         }
     }
@@ -64,10 +63,10 @@ public class DefaultSchedulerTests
     [Fact]
     public async Task WaitAsync_with_timeout()
     {
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var task = Task.Delay(TimeSpan.FromMilliseconds(10));
 
-        var result = sut.WaitAsync(task, TimeSpan.FromSeconds(1));
+        var result = task.WaitAsync(TimeSpan.FromSeconds(1), sut);
 
         await result
             .Should()
@@ -77,10 +76,10 @@ public class DefaultSchedulerTests
     [Fact]
     public async Task WaitAsync_of_T_with_timeout()
     {
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var task = StringTask();
 
-        var result = sut.WaitAsync(task, TimeSpan.FromSeconds(1));
+        var result = task.WaitAsync(TimeSpan.FromSeconds(1), sut);
 
         await result
             .Should()
@@ -97,10 +96,10 @@ public class DefaultSchedulerTests
     public async Task WaitAsync_with_timeout_and_cancellationToken()
     {
         using var cts = new CancellationTokenSource();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var task = Task.Delay(TimeSpan.FromMilliseconds(10));
 
-        var result = sut.WaitAsync(task, TimeSpan.FromSeconds(1), cts.Token);
+        var result = task.WaitAsync(TimeSpan.FromSeconds(1), sut, cts.Token);
 
         await result
             .Should()
@@ -111,10 +110,10 @@ public class DefaultSchedulerTests
     public async Task WaitAsync_of_T_with_timeout_and_cancellationToken()
     {
         using var cts = new CancellationTokenSource();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var task = StringTask();
 
-        var result = sut.WaitAsync(task, TimeSpan.FromSeconds(1), cts.Token);
+        var result = task.WaitAsync(TimeSpan.FromSeconds(1), sut, cts.Token);
 
         await result
             .Should()
@@ -130,10 +129,10 @@ public class DefaultSchedulerTests
     [Fact]
     public async Task WaitAsync_with_timeout_throws()
     {
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var task = Task.Delay(TimeSpan.FromSeconds(1));
 
-        var result = sut.WaitAsync(task, TimeSpan.FromMilliseconds(5));
+        var result = task.WaitAsync(TimeSpan.FromMilliseconds(5), sut);
 
         await result
             .Awaiting(x => x)
@@ -144,10 +143,10 @@ public class DefaultSchedulerTests
     [Fact]
     public async Task WaitAsync_of_T_with_timeout_throws()
     {
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var task = StringTask();
 
-        var result = sut.WaitAsync(task, TimeSpan.FromMilliseconds(5));
+        var result = task.WaitAsync(TimeSpan.FromMilliseconds(5), sut);
 
         await result
             .Awaiting(x => x)
@@ -165,10 +164,10 @@ public class DefaultSchedulerTests
     public async Task WaitAsync_with_timeout_and_cancellationToken_throws()
     {
         using var cts = new CancellationTokenSource();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var task = Task.Delay(TimeSpan.FromSeconds(1));
 
-        var result = sut.WaitAsync(task, TimeSpan.FromSeconds(2), cts.Token);
+        var result = task.WaitAsync(TimeSpan.FromSeconds(2), sut, cts.Token);
         cts.Cancel();
 
         await result
@@ -181,10 +180,10 @@ public class DefaultSchedulerTests
     public async Task WaitAsync_of_T_with_timeout_and_cancellationToken_throws()
     {
         using var cts = new CancellationTokenSource();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var task = StringTask();
 
-        var result = sut.WaitAsync(task, TimeSpan.FromSeconds(2), cts.Token);
+        var result = task.WaitAsync(TimeSpan.FromSeconds(2), sut, cts.Token);
         cts.Cancel();
 
         await result
@@ -204,9 +203,9 @@ public class DefaultSchedulerTests
     {
         var cts = new CancellationTokenSource();
         cts.Dispose();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
 
-        var throws = () => sut.CancelAfter(cts, TimeSpan.Zero);
+        var throws = () => cts.CancelAfter(TimeSpan.Zero, sut);
 
         throws.Should().ThrowExactly<ObjectDisposedException>();
     }
@@ -217,9 +216,9 @@ public class DefaultSchedulerTests
     public void CancelAfter_throws_ArgumentOutOfRangeException(double timespanInMilliseconds)
     {
         using var cts = new CancellationTokenSource();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
 
-        var throws = () => sut.CancelAfter(cts, TimeSpan.FromMilliseconds(timespanInMilliseconds));
+        var throws = () => cts.CancelAfter(TimeSpan.FromMilliseconds(timespanInMilliseconds), sut);
 
         throws.Should().ThrowExactly<ArgumentOutOfRangeException>();
     }
@@ -227,9 +226,9 @@ public class DefaultSchedulerTests
     [Fact]
     public void CancelAfter_throws_ArgumentNullException()
     {
-        var sut = DefaultScheduler.Instance;
+        using var cts = new CancellationTokenSource();
 
-        var throws = () => sut.CancelAfter(default!, TimeSpan.Zero);
+        var throws = () => cts.CancelAfter(TimeSpan.Zero, default(TimeProvider)!);
 
         throws.Should().ThrowExactly<ArgumentNullException>();
     }
@@ -238,9 +237,9 @@ public class DefaultSchedulerTests
     public async Task CancelAfter_cancels_delay_eq_zero()
     {
         using var cts = new CancellationTokenSource();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
 
-        sut.CancelAfter(cts, TimeSpan.Zero);
+        cts.CancelAfter(TimeSpan.Zero, sut);
 
         await sut.Delay(TimeSpan.FromMilliseconds(50));
         cts.IsCancellationRequested.Should().BeTrue();
@@ -250,10 +249,10 @@ public class DefaultSchedulerTests
     public async Task CancelAfter_cancels()
     {
         using var cts = new CancellationTokenSource();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var delay = TimeSpan.FromMilliseconds(30);
 
-        sut.CancelAfter(cts, delay);
+        cts.CancelAfter(delay, sut);
 
         var throwsAfterCancel = () => Task.Delay(TimeSpan.FromHours(1), cts.Token);
         await throwsAfterCancel.Should().ThrowAsync<TaskCanceledException>();
@@ -263,15 +262,15 @@ public class DefaultSchedulerTests
     public async Task CancelAfter_reschedule_cancel()
     {
         using var cts = new CancellationTokenSource();
-        var sut = DefaultScheduler.Instance;
+        var sut = TimeProvider.System;
         var delay1 = TimeSpan.FromSeconds(1);
         var delay2 = TimeSpan.FromMilliseconds(50);
 
-        sut.CancelAfter(cts, delay1);
-        sut.CancelAfter(cts, delay2);
+        cts.CancelAfter(delay1, sut);
+        cts.CancelAfter(delay2, sut);
 
         // add extra buffer to ensure cancellation has been processed
         await sut.Delay(delay2 * 2);
         cts.IsCancellationRequested.Should().BeTrue();
-    }  
+    }
 }