+++ test/TimeScheduler.Tests/Testing/TestSchedulerPeriodicTimerTests.cs
@@ -1,5 +1,6 @@
 namespace TimeScheduler.Testing;
 
+[Obsolete] // marked obsolete to stop warnings related to SUT
 public class TestSchedulerPeriodicTimerTests
 {    
     [Fact]