+++ test/TimeScheduler.Tests/Testing/TestSchedulerWaitAsyncTests.cs
@@ -1,8 +1,6 @@
-using System.Threading.Tasks;
-using FluentAssertions;
-
 namespace TimeScheduler.Testing;
 
+[Obsolete] // marked obsolete to stop warnings related to SUT
 public class TestSchedulerWaitAsyncTests
 {
     internal const uint MaxSupportedTimeout = 0xfffffffe;